+++ R/cleave.R
@@ -1,19 +1,22 @@
 #' Cleave
 #' 
-#' Customise error messages.
+#' Customise hard error messages.
 #' 
 #' @param html Html content to display instead of error messages, 
-#' if \code{NULL} the original message is displyed.
+#' if `NULL` the original message is displyed.
 #' @param color Color of error message text.
 #' @param bg_color Background color of error message overlay. 
-#' If \code{NULL} will be transparent.
+#' If `NULL` will be transparent.
 #' @param center_vertical,center_horizontal Whether to center the message vertically and horizontally, a \code{boolean}.
 #' @param session A valid shiny session.
-#' @param ids Ids of elements to apply the cleave to. If \code{NULL} applies to all error messages.
+#' @param ids Ids of elements to apply the cleave to. If `NULL` applies to all error messages.
 #' @param duration Duration of animation showing the error message, set to \code{0} to have none.
-#' @param silent_errors Set to TRUE to also cleave silent errors: errors that are raised by \code{shiny::validate}.
+#' @param silent_errors Set to TRUE to also cleave silent errors: errors that are raised by [shiny::validate()].
+#' Alternatively you might want to use [chisel()] to specifically stylise those messages.
 #' 
 #' @return None
+#' 
+#' @seealso [chisel()] for soft error messages.
 #' 
 #' @export
 cleave <- function(html = NULL, color = NULL, bg_color = NULL, duration = NULL, center_vertical = NULL,
@@ -58,7 +61,7 @@ cleave <- function(html = NULL, color = NULL, bg_color = NULL, duration = NULL, 
 
 #' Cleave Theme
 #' 
-#' Define a theme to apply to all subsequent \code{\link{cleave}}.
+#' Define a theme to apply to all subsequent [cleave()].
 #' 
 #' @inheritParams cleave
 #'