+++ docs/docs/changelog.md
@@ -4,6 +4,7 @@
 
 * Fix passing multiple ids to `cleave` see [#1](https://github.com/JohnCoene/sever/issues/1)
 * Fix to remove cleave error on new value render + empty HTML content on error, see [#2](https://github.com/JohnCoene/sever/issues/2).
+* Add `chisel` to style soft error messages, those raised by `validate`.
 
 ## sever 0.0.4